fix: enable error output in received callback
Previously there was an issue with errors draining memory and consuming
a large amount of CPU time.
This was due to object inspection of both Q and Loop classes which
cyclicly referenced each other in many cases. By overwriting the
default inspect methods to only output limited and far more useful
information this issue was mitigated and we can re-enable dynamic error
inspection